Add clone() method
Recursive layers were overseen. They now throw a TypeError at `renderLayer()` when a cyclic relation has been found, and a new method has been added on the CanvasLayer prototype to allow the cloning of the CanvasLayer. This method returns a new CanvasLayer with all the saved commands from the original CanvasLayer, actually creating a fork from the original CanvasLayer instance.